HEVM: split utils files
